revised version of hrnet and shufflenet
