+++ hub_module/modules/image/semantic_segmentation/humanseg_lite/README.md
@@ -1,26 +1,21 @@
-## 模型概述
-
-HUmanSeg_lite是在ShuffleNetV2网络结构的基础上进行优化，进一步减小了网络规模，网络大小只有541K，量化后只有187K， 适用于手机自拍人像分割，且能在移动端进行实时分割。
-
-
 ## 命令行预测
 
 **cpu运行命令**
 ```
-hub run shufflenet_humanseg --input_path test.txt --visualization True --batch_size 2
+hub run humanseg_lite --input_path path.txt --visualization True --batch_size 2
 
 ```
 **gpu运行命令**
 
 ```
-CUDA_VISIBLE_DEVICES=0 hub run shufflenet_humanseg --input_path test.txt --visualization True --batch_size 2 --use_gpu True
+CUDA_VISIBLE_DEVICES=0 hub run humanseg_lite --input_path path.txt --visualization True --batch_size 2 --use_gpu True
 
 ```
 
 ## API
 
 ```python
-def segmentation(images=None,
+def segment(images=None,
                  paths=None,
                  batch_size=1,
                  use_gpu=False,
@@ -33,7 +28,7 @@ def segmentation(images=None,
 **参数**
 
 * images (list\[numpy.ndarray\]): 图片数据，ndarray.shape 为 \[H, W, C\]，BGR格式；
-* paths (list\[str\]): 图片的路径；
+* paths (txt文件): 保存图片的路径的txt文件；
 * batch\_size (int): batch 的大小；
 * use\_gpu (bool): 是否使用 GPU；
 * visualization (bool): 是否将识别结果保存为图片文件；
@@ -67,9 +62,9 @@ def save_inference_model(dirname,
 import cv2
 import paddlehub as hub
 
-classifier = hub.Module('shufflenet_humanseg')
+classifier = hub.Module('humanseg_lite')
 im = cv2.imread('检测照片.jpg')
-res = classifier.segmentation(images=[im],visualization=True)
+res = classifier.segment(images=[im],visualization=True)
 ```
 
 ## 服务部署
@@ -80,7 +75,7 @@ PaddleHub Serving可以部署一个人像分割的在线服务。
 
 运行启动命令：
 ```shell
-$ hub serving start -m shufflenet_humanseg
+$ hub serving start -m humanseg_lite
 ```
 
 这样就完成了一个人像分割的服务化API的部署，默认端口号为8866。
@@ -111,7 +106,7 @@ def base64_to_cv2(b64str):
 # 发送HTTP请求
 data = {'images':[cv2_to_base64(cv2.imread("检测照片.jpg"))]}
 headers = {"Content-type": "application/json"}
-url = "http://127.0.0.1:8866/predict/shufflenet_humanseg"
+url = "http://127.0.0.1:8866/predict/humanseg_lite"
 r = requests.post(url=url, headers=headers, data=json.dumps(data))
 
 # 打印预测结果
