+++ hub_module/tests/unittests/test_humanseg_lite.py
@@ -20,7 +20,7 @@ class TestHumanSeg(unittest.TestCase):
     @classmethod
     def setUpClass(self):
         """Prepare the environment once before execution of all tests.\n"""
-        self.human_seg = hub.Module(name="hrnet_w18_samll_v1_humanseg")
+        self.human_seg = hub.Module(name="humanseg_lite")
 
     @classmethod
     def tearDownClass(self):
@@ -48,10 +48,13 @@ class TestHumanSeg(unittest.TestCase):
 
     def test_batch(self):
         with fluid.program_guard(self.test_prog):
+            pics_path_list = [
+                os.path.join(pic_dir, f) for f in os.listdir(pic_dir)
+            ]
             result = self.human_seg.segment(
                 paths=path_txt,
-                batch_size=5,
-                output_dir='batch_output_hrnet',
+                batch_size=2,
+                output_dir='batch_output_shuffle',
                 use_gpu=True,
                 visualization=True)
             print(result)
@@ -63,16 +66,17 @@ class TestHumanSeg(unittest.TestCase):
             ]
             pics_ndarray = list()
             for pic_path in pics_path_list:
+                img = cv2.imread(pic_path)
                 result = self.human_seg.segment(
-                    images=[cv2.imread(pic_path)],
-                    output_dir='ndarray_output_hrnet',
+                    images=[img],
+                    output_dir='ndarray_output_shuffle',
                     use_gpu=True,
                     visualization=True)
 
     def test_save_inference_model(self):
         with fluid.program_guard(self.test_prog):
             self.human_seg.save_inference_model(
-                dirname='hrnet_w18_samll_v1_humanseg', combined=True)
+                dirname='humanseg_lite', combined=True)
 
 
 if __name__ == "__main__":
