+++ hub_module/tests/unittests/test_humanseg_mobile.py
@@ -20,7 +20,7 @@ class TestHumanSeg(unittest.TestCase):
     @classmethod
     def setUpClass(self):
         """Prepare the environment once before execution of all tests.\n"""
-        self.human_seg = hub.Module(name="shufflenet_humanseg")
+        self.human_seg = hub.Module(name="humanseg_mobile")
 
     @classmethod
     def tearDownClass(self):
@@ -43,19 +43,16 @@ class TestHumanSeg(unittest.TestCase):
 
             img = cv2.imread(pics_path_list[0])
             result = self.human_seg.segment(
-                images=[img], use_gpu=True, visualization=True)
+                images=[img], use_gpu=False, visualization=True)
             print(result[0]['data'])
 
     def test_batch(self):
         with fluid.program_guard(self.test_prog):
-            pics_path_list = [
-                os.path.join(pic_dir, f) for f in os.listdir(pic_dir)
-            ]
             result = self.human_seg.segment(
                 paths=path_txt,
-                batch_size=2,
-                output_dir='batch_output_shuffle',
-                use_gpu=True,
+                batch_size=5,
+                output_dir='batch_output_hrnet',
+                use_gpu=False,
                 visualization=True)
             print(result)
 
@@ -66,17 +63,16 @@ class TestHumanSeg(unittest.TestCase):
             ]
             pics_ndarray = list()
             for pic_path in pics_path_list:
-                img = cv2.imread(pic_path)
                 result = self.human_seg.segment(
-                    images=[img],
-                    output_dir='ndarray_output_shuffle',
-                    use_gpu=True,
+                    images=[cv2.imread(pic_path)],
+                    output_dir='ndarray_output_hrnet',
+                    use_gpu=False,
                     visualization=True)
 
     def test_save_inference_model(self):
         with fluid.program_guard(self.test_prog):
             self.human_seg.save_inference_model(
-                dirname='shufflenet_humanseg', combined=True)
+                dirname='humanseg_mobile', combined=True)
 
 
 if __name__ == "__main__":
